Remove the now-unused ComposerSource enum
